Nightly Build 17 Mar 09
Almost Version .7. The scroll zoom functionality now works on the main image (in tile view, need to add support for smaller images) and thumbnail (all the time). In addition, when clicking on a zoom in/zoom out button, the program zooms into the middle of the viewer.

Also added in cool cursors (basically yet another ripoff of Google Maps). I took the cursors from Trolltech's Qt website, and may need to give cred.